Add game window position/size to viewport configuration
On multi display systems it is useful to be able to position the game
window on specific displays. Add new config variables in order to not
mix the window positioning with the viewport (i.e. the part of the
window being rendered to).
This does change the configuration format for viewports:
x=N, y=M -> loc="X=N,Y=M"
width=N, height=M -> size="X=N,Y=M"